+++ PYTHON/Primeiro-site-flask/flaskimpressionador/create_db.py
@@ -1,0 +1,5 @@
+from flaskimpressionador import app, database
+
+with app.app_context():
+    database.create_all()
+    print("Banco de dados criado com sucesso!")
