+++ PYTHON/Primeiro-site-flask/flaskimpressionador/models.py
@@ -1,0 +1,19 @@
+from datetime import datetime
+from flaskimpressionador import database
+
+class Usuario(database.Model):
+    id= database.Column(database.Integer, primary_key=True)
+    username= database.Column(database.String(120), nullable=False)
+    email= database.Column(database.String(120), nullable=False,unique=True)
+    senha= database.Column(database.String(120), nullable=False)
+    foto_perfil= database.Column(database.String,default='default.jpg')
+    cursos = database.Column(database.String, nullable=False, default='Não informado')
+    posts= database.relationship('Post', backref='autor_post', lazy=True)
+
+
+class Post(database.Model):
+    id = database.Column(database.Integer, primary_key=True)
+    titulo = database.Column(database.String(120), nullable=False)
+    corpo = database.Column(database.Text, nullable=False)
+    data_criacao= database.Column(database.DateTime, nullable=False, default=datetime.utcnow)
+    id_usuario= database.Column(database.Integer, database.ForeignKey('usuario.id'),nullable=False)
